+++ NEWS.md
@@ -28,5 +28,6 @@
 - Default countdown value depends on if the checkbox is initially checked or not.
 - Fixed bug when showing "MEASURING...", "MEASURE NOW" message.
 - Fixed error in asd_acquire callback in Cross/Mesh
+- Fixed the asd data printing format.
 
 ## News